angaza_ecommerce/requirements.txt
asgiref==3.3.4
certifi==2021.5.30
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.3
django-admin-rangefilter==0.8.1
django-cors-headers==3.7.0
django-filter==2.4.0
django-import-export==2.5.0
django-jsonfield==1.4.1
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
MarkupPy==1.14
MarkupSafe==2.0.1
mysqlclient==2.0.3
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.7
Pillow==8.2.0
psycopg2==2.9.1
PyJWT==1.7.1
pyotp==2.6.0
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
validate-email==1.3
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
